# contracts/generated/api-v2.openapi.json
{
  "components": {
    "schemas": {
...
      "CommandResult_RunSnapshotV1": {
        "additionalProperties": false,
        "properties": {
          "replayed": {
            "type": "boolean"
          },
          "resource": {
            "additionalProperties": false,
            "properties": {
              "active_attempt_id": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/Id"
                  }
                ]
              },
              "cancellation_requested_at": {
                "format": "date-time",
                "type": [
                  "string",
                  "null"
                ]
              },
              "current_attempt_no": {
                "format": "int64",
                "maximum": 4294967295,
                "minimum": 0,
                "type": "integer"
              },
              "cycle_id": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/Id"
                  }
                ]
              },
              "deadline_at": {
                "format": "date-time",
                "type": "string"
              },
              "finished_at": {
                "format": "date-time",
                "type": [
                  "string",
                  "null"
                ]
              },
              "id": {
                "$ref": "#/components/schemas/Id"
              },
              "input_set_id": {
                "$ref": "#/components/schemas/Id"
              },
              "kind": {
                "$ref": "#/components/schemas/RunKind"
              },
              "last_event_seq": {
                "$ref": "#/components/schemas/DbCounter"
              },
              "project_id": {
                "$ref": "#/components/schemas/Id"
              },
              "queued_at": {
                "format": "date-time",
                "type": "string"
              },
              "revision": {
                "$ref": "#/components/schemas/Revision"
              },
              "schema_version": {
                "$ref": "#/components/schemas/SchemaV1"
              },
              "started_at": {
                "format": "date-time",
                "type": [
                  "string",
                  "null"
                ]
              },
              "state": {
                "$ref": "#/components/schemas/RunState"
              },
              "terminal_reason_code": {
                "type": [
                  "string",
                  "null"
                ]
              }
            },
            "required": [
              "schema_version",
              "id",
              "project_id",
              "kind",
              "input_set_id",
              "state",
              "current_attempt_no",
              "last_event_seq",
              "deadline_at",
              "queued_at",
              "revision"
            ],
            "type": "object"
          },
          "schema_version": {
            "$ref": "#/components/schemas/SchemaV1"
          }
        },
        "required": [
          "schema_version",
          "replayed",
          "resource"
        ],
        "type": "object"
      },
...
      "DbCounter": {
        "description": "Canonical decimal string in the PostgreSQL signed bigint range; nonnegative counters or positive revisions.",
        "maxLength": 19,
        "minLength": 1,
        "pattern": "^(?:[1-9][0-9]{0,17}|0|[1-8][0-9]{18}|9[0-1][0-9]{17}|92[0-1][0-9]{16}|922[0-2][0-9]{15}|9223[0-2][0-9]{14}|92233[0-6][0-9]{13}|922337[0-1][0-9]{12}|92233720[0-2][0-9]{10}|922337203[0-5][0-9]{9}|9223372036[0-7][0-9]{8}|92233720368[0-4][0-9]{7}|922337203685[0-3][0-9]{6}|9223372036854[0-6][0-9]{5}|92233720368547[0-6][0-9]{4}|922337203685477[0-4][0-9]{3}|9223372036854775[0-7][0-9]{2}|922337203685477580[0-6][0-9]{0}|9223372036854775807)(?![\\s\\S])",
        "type": "string"
      },
...
      "Id": {
        "format": "uuid",
        "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-7[0-9a-fA-F]{3}-[89aAbB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$",
        "type": "string"
      },
...
      "Page_RunSnapshotV1": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "additionalProperties": false,
              "properties": {
                "active_attempt_id": {
                  "oneOf": [
                    {
                      "type": "null"
                    },
                    {
                      "$ref": "#/components/schemas/Id"
                    }
                  ]
                },
                "cancellation_requested_at": {
                  "format": "date-time",
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "current_attempt_no": {
                  "format": "int64",
                  "maximum": 4294967295,
                  "minimum": 0,
                  "type": "integer"
                },
                "cycle_id": {
                  "oneOf": [
                    {
                      "type": "null"
                    },
                    {
                      "$ref": "#/components/schemas/Id"
                    }
                  ]
                },
                "deadline_at": {
                  "format": "date-time",
                  "type": "string"
                },
                "finished_at": {
                  "format": "date-time",
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "id": {
                  "$ref": "#/components/schemas/Id"
                },
                "input_set_id": {
                  "$ref": "#/components/schemas/Id"
                },
                "kind": {
                  "$ref": "#/components/schemas/RunKind"
                },
                "last_event_seq": {
                  "$ref": "#/components/schemas/DbCounter"
                },
                "project_id": {
                  "$ref": "#/components/schemas/Id"
                },
                "queued_at": {
                  "format": "date-time",
                  "type": "string"
                },
                "revision": {
                  "$ref": "#/components/schemas/Revision"
                },
                "schema_version": {
                  "$ref": "#/components/schemas/SchemaV1"
                },
                "started_at": {
                  "format": "date-time",
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "state": {
                  "$ref": "#/components/schemas/RunState"
                },
                "terminal_reason_code": {
                  "type": [
                    "string",
                    "null"
                  ]
                }
              },
              "required": [
                "schema_version",
                "id",
                "project_id",
                "kind",
                "input_set_id",
                "state",
                "current_attempt_no",
                "last_event_seq",
                "deadline_at",
                "queued_at",
                "revision"
              ],
              "type": "object"
            },
            "type": "array"
          },
          "next_cursor": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Id"
              }
            ]
          },
          "schema_version": {
            "$ref": "#/components/schemas/SchemaV1"
          }
        },
        "required": [
          "schema_version",
          "items"
        ],
        "type": "object"
      },
...
      "Problem": {
        "properties": {
          "code": {
            "type": "string"
          },
          "current_revision": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Revision"
              }
            ]
          },
          "detail": {
            "type": "string"
          },
          "field_errors": {
            "items": {
              "$ref": "#/components/schemas/FieldError"
            },
            "type": "array"
          },
          "request_id": {
            "$ref": "#/components/schemas/Id"
          },
          "retryable": {
            "type": "boolean"
          },
          "safe_next_actions": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "status": {
            "format": "int32",
            "maximum": 599,
            "minimum": 100,
            "type": "integer"
          },
          "title": {
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "code",
          "detail",
          "request_id",
          "retryable",
          "field_errors",
          "safe_next_actions"
        ],
        "type": "object"
      },
...
      "Revision": {
        "description": "Canonical decimal string in the PostgreSQL signed bigint range; nonnegative counters or positive revisions.",
        "maxLength": 19,
        "minLength": 1,
        "pattern": "^(?:[1-9][0-9]{0,17}|[1-8][0-9]{18}|9[0-1][0-9]{17}|92[0-1][0-9]{16}|922[0-2][0-9]{15}|9223[0-2][0-9]{14}|92233[0-6][0-9]{13}|922337[0-1][0-9]{12}|92233720[0-2][0-9]{10}|922337203[0-5][0-9]{9}|9223372036[0-7][0-9]{8}|92233720368[0-4][0-9]{7}|922337203685[0-3][0-9]{6}|9223372036854[0-6][0-9]{5}|92233720368547[0-6][0-9]{4}|922337203685477[0-4][0-9]{3}|9223372036854775[0-7][0-9]{2}|922337203685477580[0-6][0-9]{0}|9223372036854775807)(?![\\s\\S])",
        "type": "string"
      },
      "RunCancelV1": {
        "additionalProperties": false,
        "properties": {
          "expected_revision": {
            "$ref": "#/components/schemas/Revision"
          },
          "schema_version": {
            "$ref": "#/components/schemas/SchemaV1"
          }
        },
        "required": [
          "schema_version",
          "expected_revision"
        ],
        "type": "object"
      },
      "RunEventV1": {
        "properties": {
          "attempt_id": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Id"
              }
            ]
          },
          "event_type": {
            "maxLength": 120,
            "minLength": 1,
            "pattern": "^[a-z][a-z0-9_.]*(?![\\s\\S])",
            "type": "string"
          },
          "occurred_at": {
            "format": "date-time",
            "type": "string"
          },
          "payload": {
            "additionalProperties": true,
            "description": "Public extensible schema-v1 JSON object; serialized UTF-8 is limited to 65536 bytes. Known event types additionally validate their specific payload contract.",
            "properties": {
              "schema_version": {
                "enum": [
                  1
                ],
                "maximum": 1,
                "minimum": 1,
                "type": "integer"
              }
            },
            "required": [
              "schema_version"
            ],
            "type": "object"
          },
          "run_id": {
            "$ref": "#/components/schemas/Id"
          },
          "schema_version": {
            "$ref": "#/components/schemas/SchemaV1"
          },
          "seq": {
            "$ref": "#/components/schemas/DbCounter"
          }
        },
        "required": [
          "schema_version",
          "run_id",
          "seq",
          "event_type",
          "occurred_at",
          "payload"
        ],
        "type": "object"
      },
      "RunKind": {
        "enum": [
          "AGENT_RESEARCH",
          "DATA_VALIDATE",
          "ALPHA_EVALUATE",
          "PORTFOLIO_BUILD",
          "PORTFOLIO_SIMULATE",
          "FORWARD_EVALUATE",
          "EXPORT",
          "IMPORT"
        ],
        "type": "string"
      },
      "RunSnapshotV1": {
        "additionalProperties": false,
        "properties": {
          "active_attempt_id": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Id"
              }
            ]
          },
          "cancellation_requested_at": {
            "format": "date-time",
            "type": [
              "string",
              "null"
            ]
          },
          "current_attempt_no": {
            "format": "int64",
            "maximum": 4294967295,
            "minimum": 0,
            "type": "integer"
          },
          "cycle_id": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Id"
              }
            ]
          },
          "deadline_at": {
            "format": "date-time",
            "type": "string"
          },
          "finished_at": {
            "format": "date-time",
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "$ref": "#/components/schemas/Id"
          },
          "input_set_id": {
            "$ref": "#/components/schemas/Id"
          },
          "kind": {
            "$ref": "#/components/schemas/RunKind"
          },
          "last_event_seq": {
            "$ref": "#/components/schemas/DbCounter"
          },
          "project_id": {
            "$ref": "#/components/schemas/Id"
          },
          "queued_at": {
            "format": "date-time",
            "type": "string"
          },
          "revision": {
            "$ref": "#/components/schemas/Revision"
          },
          "schema_version": {
            "$ref": "#/components/schemas/SchemaV1"
          },
          "started_at": {
            "format": "date-time",
            "type": [
              "string",
              "null"
            ]
          },
          "state": {
            "$ref": "#/components/schemas/RunState"
          },
          "terminal_reason_code": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "schema_version",
          "id",
          "project_id",
          "kind",
          "input_set_id",
          "state",
          "current_attempt_no",
          "last_event_seq",
          "deadline_at",
          "queued_at",
          "revision"
        ],
        "type": "object"
      },
      "RunState": {
        "enum": [
          "QUEUED",
          "DISPATCHING",
          "RUNNING",
          "RECONCILING",
          "CANCEL_REQUESTED",
          "SUCCEEDED",
          "FAILED",
          "CANCELLED"
        ],
        "type": "string"
      },
      "SchemaV1": {
        "enum": [
          1
        ],
        "maximum": 1,
        "minimum": 1,
        "type": "integer"
      },
...
    },
    "securitySchemes": {
      "BrowserSession": {
        "description": "Native private cookie; browser writes require exact same-origin Origin. Explicit loopback development uses quazonai-dev.",
        "in": "cookie",
        "name": "__Host-quazonai",
        "type": "apiKey"
      },
      "MachineBearer": {
        "bearerFormat": "qz2.UUIDv7.opaque-capability",
        "description": "Opaque native capability; only project/run/downstream-scoped server records confer authority. Never combine with browser Cookie.",
        "scheme": "bearer",
        "type": "http"
      },
      "OperatorCommandGrant": {
        "description": "One-time TOTP-verified CLI grant bound to this credential, exact operation, target and full nonsecret request. No Agent/automation grant issuance.",
        "in": "header",
        "name": "X-Operator-Grant",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "",
    "license": {
      "name": ""
    },
    "title": "QuaZonai HTTP API",
    "version": "2.0.0-dev.1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v2/runs": {
      "get": {
        "operationId": "list_runs",
        "parameters": [
          {
            "in": "query",
            "name": "project_id",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/Id"
            }
          },
          {
            "in": "query",
            "name": "state",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/RunState"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/Id"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "format": "int32",
              "maximum": 100,
              "minimum": 1,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_RunSnapshotV1"
                }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": ""
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": ""
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": ""
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Shared native machine attempt limit or bounded crypto capacity; respect Retry-After.",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "BrowserSession": []
          },
          {
            "MachineBearer": []
          }
        ],
        "tags": [
          "Runs"
        ]
      }
    },
    "/api/v2/runs/{id}": {
      "get": {
        "operationId": "get_run",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunSnapshotV1"
                }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": ""
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": ""
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Shared native machine attempt limit or bounded crypto capacity; respect Retry-After.",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "BrowserSession": []
          },
          {
            "MachineBearer": []
          }
        ],
        "tags": [
          "Runs"
        ]
      }
    },
    "/api/v2/runs/{id}/cancel": {
      "post": {
        "operationId": "cancel",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Id"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunCancelV1"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandResult_RunSnapshotV1"
                }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": ""
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": ""
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": ""
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": ""
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Shared native machine attempt limit or bounded crypto capacity; respect Retry-After.",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "BrowserSession": []
          },
          {
            "MachineBearer": []
          }
        ],
        "tags": [
          "Runs"
        ]
      }
    },
    "/api/v2/runs/{id}/events": {
      "get": {
        "operationId": "events",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Id"
            }
          },
          {
            "description": "Exact run UUID, colon, canonical decimal sequence; omitted starts at zero.",
            "in": "header",
            "name": "Last-Event-ID",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/RunEventV1"
                }
              }
            },
            "description": "Durable RunEventV1 frames. Disconnect never cancels a run. Reconnect with the last emitted event ID."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": ""
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": ""
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": ""
          },
          "410": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": ""
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": ""
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Shared native machine attempt limit or bounded crypto capacity; respect Retry-After.",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "BrowserSession": []
          },
          {
            "MachineBearer": []
          }
        ],
        "tags": [
          "Runs"
        ]
      }
    }
  },
...
}